Backend Returns Proper String Output
There was no route defined for the root path "/", hence the build-backend job could not find the "Backend running on http://localhost:8000" etc. This should be fixed as index.js now outputs "Backend running on http://localhost:8000" to the root path. Rather than just printing it, it sends it to the client.

This should've been here from the beginning apologies.