Fixed Split Second Error Popup
Enrico said he found this during testing: "After signing in, a message comes up letting you know you have signed up, but for a split second, it said error (not really sure why it happened as I successfully logged on to the database)." - this is now fixed. This happened because both alerts shared the same Snackbar component. So I separated the components ez

minor change i added within signup is adding the loading circle and only redirecting the user once that loading circle is done (2 seconds long), to let the alert pop up to let them know that they've signed up successfully or not